const inputs are now longs

+++ backend/tests/test-clear-uint8_t-addConst.cpp
@@ -27,7 +27,7 @@ int main(void) {
   // inputs is vector of vectors
   std::vector<std::vector<uint8_t>> inputs = {{117, 2, 3, 4}};
   // const_inputs is vector (same across slots)
-  std::vector<uint8_t> const_inputs = {6};
+  std::vector<long> const_inputs = {6};
   std::vector<uint8_t> exp_values = {123, 8, 9, 10};
 
   std::vector<std::vector<uint8_t>> result =
